refactor code; add policy; add service;
