+++ aopt/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aopt"
-version = "0.4.4"
+version = "0.5.0"
 authors = ["araraloren <blackcatoverwall@gmail.com>"]
 edition = "2021"
 license = "MPL-2.0"
