+++ aopt/src/proc.rs
@@ -5,6 +5,7 @@ use std::fmt::Debug;
 
 use crate::ctx::Context;
 use crate::err::Result;
+use crate::parser::ValueKeeper;
 use crate::set::Set;
 use crate::uid::Uid;
 
@@ -15,11 +16,23 @@ pub trait Info: Debug {
     fn info_uid(&self) -> Uid;
 }
 
-pub trait Proc<M: Matcher>: Debug {
-    fn process(&mut self, matcher: &mut M, set: &mut dyn Set) -> Result<bool>;
+pub trait Matching<M: Matcher>: Debug {
+    fn matching_opt<S: Set>(
+        &mut self,
+        matcher: &mut M,
+        set: &mut S,
+        invoke: bool,
+    ) -> Result<Vec<ValueKeeper>>;
+
+    fn matching_nonopt<S: Set>(
+        &mut self,
+        matcher: &mut M,
+        set: &mut S,
+        invoke: bool,
+    ) -> Result<Vec<ValueKeeper>>;
 }
 
-pub trait Matcher: Debug {
+pub trait Matcher: Debug + Default {
     fn uid(&self) -> Uid;
 
     fn add_ctx(&mut self, ctx: Box<dyn Context>);
