+++ snowball-follow/Cargo.toml
@@ -10,8 +10,8 @@ categories = ["command-line-interface"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-aopt = { path = "../aopt", version = "0.4.3" }
-aopt-help = { path = "../aopt-help", version = "0.2.5" }
+aopt = "0.4.3"
+aopt-help = "0.2.5"
 tracing-subscriber = { version = "0.3.2", features = ["env-filter"] }
 color-eyre = "0.5.11"
 reqwest = { version = "0.11.6", features = ["json", "gzip", "cookies"] }
